// admin-panel/src/app/@shared/enums/permissions.enum.ts
export enum PermissionsEnum {
  USER_VIEW = 'UserView',
  USER_MODIFY = 'UserModify',
  USER_DELETE = 'UserDelete',
  USER_CREATE = 'UserCreate',
  PERMISSION_VIEW = 'PermissionView',
  PERMISSION_MODIFY = 'PermissionModify',
  PERMISSION_DELETE = 'PermissionDelete',
  PERMISSION_CREATE = 'PermissionCreate',
  ROLE_CREATE = 'RoleCreate',
  ROLE_MODIFY = 'RoleModify',
  ROLE_VIEW = 'RoleView',
  ROLE_DELETE = 'RoleDelete',
  PERMISSION_TYPE_DELETE = 'PermissionTypeDelete',
  PERMISSION_TYPE_VIEW = 'PermissionTypeView',
  PERMISSION_TYPE_MODIFY = 'PermissionTypeModify',
  PERMISSION_TYPE_CREATE = 'PermissionTypeCreate',
  ROLE_PERMISSION_CREATE = 'RolePermissionCreate',
  ROLE_PERMISSION_DELETE = 'RolePermissionDelete',
  ROLE_PERMISSION_MODIFY = 'RolePermissionModify',
  ROLE_PERMISSION_VIEW = 'RolePermissionView',
}